Add files via upload
Commands:
npm install ***

Where *** is the name of the module as shown in the Dependencies. For example, to install node-cron use npm install node-cron and to install nodemailer use npm install nodemailer

Dependencies:

  "dependencies": {
    "cron": "^1.3.0",
    "es6-promise-pool": "^2.5.0",
    "firebase": "^4.6.1",
    "firebase-admin": "^5.2.1",
    "firebase-functions": "^0.6.3",
    "gmail": "^0.5.0",
    "mathjs": "^3.16.5",
    "node-cron": "^1.2.1",
    "nodemailer": "^4.3.1",
    "ping": "^0.2.2",
    "request-promise": "^4.2.2",
    "secure-compare": "^3.0.1"
  },

Description:

Takes the time stamp of the last value pushed to the database and compares with the server current time. If the difference exceeds 5 minutes send a system down email. If the system recovers from a down state send an email to indicate firebase is receiving data again. Cron is used to execute this function every 5 minutes, provided by cron-job.org

Limitations:

Cron requires the URL of the cloud function - if you re-name it the same must be applied to the cron job.